Cypress: util refactoring
