+++ cypress/support/commands/overwrite/should.ts
@@ -1,4 +1,4 @@
-import { mapAlias } from "./util";
+import { mapAlias } from "../../util";
 
 Cypress.Commands.overwrite<"should", "optional">(
   "should",
