+++ cypress/support/util.ts
@@ -1,3 +1,19 @@
-export default function getRandomText() {
+export function getRandomText() {
   return crypto.randomUUID().replace(/-/g, "").toUpperCase().substring(0, 10);
+}
+
+export function mapAlias(
+  thisObject: Record<string, unknown>,
+  argument
+): unknown {
+  if (typeof argument === "string" && argument.startsWith("@")) {
+    const alias = argument.slice(1);
+    if (Object.keys(thisObject).includes(alias)) {
+      return thisObject[alias];
+    }
+
+    throw new Error(`Alias '${argument}' was not found in test scope.`);
+  }
+
+  return argument;
 }
